+++ policyengine_us/reforms/reforms.py
@@ -140,6 +140,9 @@ from .states.va.dependent_exemption.va_dependent_exemption_reform import (
 from .states.va.hb979.va_hb979_reform import (
     create_va_hb979_reform,
 )
+from .states.ct.refundable_ctc import (
+    create_ct_refundable_ctc_reform,
+)
 from .aca import (
     create_aca_ptc_additional_bracket_reform,
     create_aca_ptc_simplified_bracket_reform,
@@ -147,6 +150,15 @@ from .aca import (
 )
 from .cdcc import (
     create_cdcc_single_parent_work_requirement_reform,
+)
+from .states.ct.sb100 import (
+    create_ct_sb100_reform,
+)
+from .states.ct.tax_rebate_2026 import (
+    create_ct_tax_rebate_2026_reform,
+)
+from .states.ct.hb5009 import (
+    create_ct_hb5009_reform,
 )
 
 
@@ -326,6 +338,7 @@ def create_structural_reforms_from_parameters(parameters, period):
         parameters, period
     )
     va_hb979 = create_va_hb979_reform(parameters, period)
+    ct_refundable_ctc = create_ct_refundable_ctc_reform(parameters, period)
     aca_ptc_additional_bracket = create_aca_ptc_additional_bracket_reform(
         parameters, period
     )
@@ -338,6 +351,9 @@ def create_structural_reforms_from_parameters(parameters, period):
     cdcc_single_parent_work_requirement = (
         create_cdcc_single_parent_work_requirement_reform(parameters, period)
     )
+    ct_sb100 = create_ct_sb100_reform(parameters, period)
+    ct_tax_rebate_2026 = create_ct_tax_rebate_2026_reform(parameters, period)
+    ct_hb5009 = create_ct_hb5009_reform(parameters, period)
 
     reforms = [
         afa_reform,
@@ -404,10 +420,14 @@ def create_structural_reforms_from_parameters(parameters, period):
         or_dependent_exemption_credit,
         va_dependent_exemption,
         va_hb979,
+        ct_refundable_ctc,
         aca_ptc_additional_bracket,
         aca_ptc_simplified_bracket,
         aca_ptc_700_fpl_cliff,
         cdcc_single_parent_work_requirement,
+        ct_hb5009,
+        ct_sb100,
+        ct_tax_rebate_2026,
     ]
     reforms = tuple(filter(lambda x: x is not None, reforms))